hide lock error on delete_metric
tooz non-blocking lock will raise an error if done through
contextmanager. delete_metric uses this functionality. we shouldn't
show an error log in this case since it likely means another process
is handling it. this raises an non-critical error to be caught so
worker skips processing, either another process handles it, or we
will come back to it.

also, metric should only be removed from indexer if storage completes
succesfully.

+++ gnocchi/storage/_carbonara.py
@@ -337,9 +337,16 @@ class CarbonaraBasedStorage(storage.StorageDriver):
 
     def delete_metric(self, metric, sync=False):
         LOG.debug("Deleting metric %s", metric)
-        with self._lock(metric.id)(blocking=sync):
+        lock = self._lock(metric.id)
+        if not lock.acquire(blocking=sync):
+            raise storage.LockedMetric(metric)
+        try:
             self._delete_metric(metric)
             self.incoming.delete_unprocessed_measures_for_metric_id(metric.id)
+        except Exception:
+            raise
+        finally:
+            lock.release()
 
     @staticmethod
     def _delete_metric_measures(metric, timestamp_key,
